requirements.txt
torch>=0.4.0
torchvision
matplotlib
numpy
scipy
pillow
urllib3
scikit-image
deeplake<4
torchmetrics
torch-fidelity